ssh: remove depdency on libsodium
This depends on a change proposed in https://nest.pijul.com/pijul/thrussh/discussions/47.
The patch allows us to avoid the dependency on `thrussh-keys` and
instead rely on a generic client. This means we have to implement some
traits for the zebra keys and signature, but they are very similar to
the implementations for the ed25519 key in thrussh-keys.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>