signer: add Signer impl for Arc
In radicle-link there are some gymnastics to allow us to use a boxed
version of a signer that is also able to be cloned.

For us to be able to use the `Signer` returned by `SshAgent::connect` as
a BoxedSigner, we would require this `Signer` to be cloned. One simple
way of doing this by having an implementation of `Signer` for `Arc<S>`
where the `S` is a `Signer`. This gives us the `Signer` and `Clone`
traits required to make a `BoxedSigner`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>