ssh: make socket path configurable
To aid testing infrastructure we want the path to the ssh-agent client
to be configurable.

To accomplish this we add an extra field to the SshAgent struct which is
the path. It is optional and so if it is not set then the SSH_AUTH_SOCK
value is used.

The add_key, remove_key, and list_keys functions now accept the SshAgent
as a parameter to keep the paths to the socket consistent.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>