meta: relax crypto dependency's semver
Pinning crypto libraries in this library makes it hard to manage
dependencies in the binary consumer side that relies on this lib.

For these reasons it is recommended in Rust idiomatic way to manage
the dependency bumps via Cargo.lock in the binary side instead as part
of regular maintenance chores.

This change both bumps up the crypto to now required mininum versions
and makes it flexible to bump up the crypto via binaries that uses
the lock file to manage dependency version bumps.

Signed-off-by: pinkforest <36498018+pinkforest@users.noreply.github.com>